refactor: fix the package structure
