+++ src/main/kotlin/lotto/domain/InputParser.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 object InputParser {
     private const val DEFAULT_DELIMITER = ","
