+++ src/main/kotlin/lotto/domain/LottoStore.kt
@@ -1,4 +1,6 @@
-package lotto
+package lotto.domain
+
+import lotto.domain.model.Lotto
 
 object LottoStore {
     fun buy(money: Int): List<Lotto> = List(money / Lotto.PRICE) { Lotto() }
