+++ src/main/kotlin/lotto/domain/model/InputResult.kt
@@ -1,3 +1,3 @@
-package lotto
+package lotto.domain.model
 
 data class InputResult(val lottos: List<Lotto>, val winningNumbers: List<Int>, val money: Int)
