+++ src/main/kotlin/lotto/domain/model/LottoResult.kt
@@ -1,3 +1,3 @@
-package lotto
+package lotto.domain.model
 
 data class LottoResult(val count: Int, val prize: Lotto.Prize)
