+++ src/main/kotlin/lotto/main.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.controller.LottoController
+
 fun main() {
     LottoController().execute()
 }
