+++ src/test/kotlin/lotto/domain/InputParserTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
